+++ gulpfile.js
@@ -14,6 +14,7 @@ const jsFiles = [
   './node_modules/chart.js/dist/Chart.js',
   './node_modules/numeral/numeral.js',
   './node_modules/underscore/underscore.js',
+  './node_modules/xlsx/dist/xlsx.full.min.js',
   './src/js/helpers.js',
   './src/js/TimeGraph.js',
   './src/js/graphSetup.js',